half  precision and missed rngs

+++ conformer/conv_subsampler.py
@@ -5,10 +5,13 @@ import jax
 
 class ConvolutionSubsampling(nnx.Module):
     def __init__(self, output_dim: int, *, rngs: nnx.Rngs):
-        self.conv1 = nnx.Conv(in_features=1, out_features=output_dim, kernel_size=(3, 3), strides=(2, 2), rngs=rngs)
-        self.conv2 = nnx.Conv(in_features=output_dim, out_features=output_dim, kernel_size=(3, 3), strides=(2, 2), rngs=rngs)
+        self.conv1 = nnx.Conv(in_features=1, out_features=output_dim, kernel_size=(3, 3),
+                            strides=(2, 2), dtype=jnp.float16, rngs=rngs)
+        self.conv2 = nnx.Conv(in_features=output_dim, out_features=output_dim, kernel_size=(3, 3),
+                            strides=(2, 2), dtype=jnp.float16, rngs=rngs)
         # D * F/4 (F = mel_bins 80, hence 20)
-        self.linear = nnx.Linear(in_features=20 * output_dim, out_features=output_dim, rngs=rngs)
+        self.linear = nnx.Linear(in_features=20 * output_dim, out_features=output_dim,
+                                  dtype=jnp.float16, rngs=rngs)
         self.dropout = nnx.Dropout(0.1, rngs=rngs)
 
     def __call__(self, x: jax.Array, *, training: bool):
